Internet Explorer does not support closest
I ran into this issue with a bootstrap render suggested for aurelia-validatejs

https://developer.mozilla.org/en-US/docs/Web/API/Element/closest

I have a proposal but there is a possible infinity loop or something. My code traverses up from the target assuming there is an element with the class form-group. There might need to be another test for hitting the body element or something in case the , then aborting. I searched for a poly fill for closest but could not find it. There might even be a find ancestors besides closest.